Added delete option to launch experiments script

+++ launch_experiments.sh
@@ -1,10 +1,34 @@
 #!/usr/bin/env bash
 
+POSITIONAL=()
+while [[ $# -gt 0 ]]
+do
+    key="$1"
+
+    case $key in
+        -d|--delete)
+        DELETE_SESSIONS=True
+        shift # past argument
+        shift # past value
+        ;;
+        *)    # unknown option
+        POSITIONAL+=("$1") # save it in an array for later
+        shift # past argument
+        ;;
+    esac
+done
+set -- "${POSITIONAL[@]}" # restore positional parameters
+
 functions=("01" "02" "03" "04" "05" "06" "07" "08" "09" "10" "11" "12" "13" "14" "15" "16" "17" "18")
 
 for i in ${functions[*]}
 do
-    echo ${i}
-    tmux kill-session -t C${i}
-    tmux new-session -d -s C${i} ./run.sh -r 25 -f C${i} -c ./config.ini -l ./logs/run2/C${i}
+    if [[ -z ${DELETE_SESSIONS+x} ]]
+    then
+        echo Creating ${i}
+        tmux new-session -d -s C${i} ./run.sh -r 25 -f C${i} -c ./config.ini -l ./logs/run2/C${i}
+    else
+        echo Deleting ${i}
+        tmux kill-session -t C${i}
+    fi
 done
